initial sketch of XML ingestion tool (#83)
Lays out the design and sets up some of the golang / CI foundation.

Use the `slog` for logging, it can do structured logging later if we
want to feed these logs to grafana. This lib is scheduled to become part
of the standard lib in future versions of golang.

Sticking with the stdlib `flag` instead of the popular `cobra` lib for
CLI parsing, I think the needs here are simple enough to get away with
it.