AMPEX-71: parse <syntax> to markdown (#91)
There's too much variety in these to encode the semantics in a sensible
way. This basically ports the
[xslt](http://hg.nginx.org/nginx.org/file/tip/xslt/content.xslt) to
generate markdown instead of HTML.

Used some scratch code in main.go to print out all the `<syntax>`
variations in use, and copied from there to determine the test cases.

Reorganized the code a little to account for similarities between
`<para>` and `<syntax>` and try to group similar structs in the same
file.