AMPEX-72: support <link> elements (#130)
* AMPEX-72: support <link> elements

Converts `<link>`s to markdown. This took a lot of restructuring and a
hack.

The main problem is that `<link>` is context sensitive:

- `<link doc="../page.xml">title</link>` links somewhere else based on
  what module file is currently being parsed
- `<link doc="page.xml">` uses the `name` attr from page.xml as the link
  title, again relative to the path of the module currently being parsed
- `<link id="anchor"/>` links to another directive on the current
  module

So to calculate the href and title of a link, we need to know:

1. the paths, links, and names of all pages we might link to:
   `<module>`s or `<article>`s
2. the link and path of the module we're currently parsing

We still want the markdown conversion to happen as part of
`xml.Unmarshal`ing because that's when we have access to the order of
XML elements, but the
[`xml.Unmarshaler`](https://pkg.go.dev/encoding/xml#Unmarshaler) has no
way to provide extra context like "other pages we might link to" or "the
current module being parsed"

The hack is to use a package variable (protected by a mutex) to give
`xml.Unmarshaler`s more context. Using the mutexes to serialize
concurrent usage seemed easier than manually tweaking the tests to
selectively remove `t.Parallel`.

Considered adding a post-processing step on the markdown using
[`text/Template`](https://pkg.go.dev/text/template), but that seemed way
more complicated than the package vars.

* optimize what files we look at