Add example ref-browser
Ports one of our useful experiments on top of the `reference-lib`, a few react components to show the directives easily.

Some of the `reference-lib` usage is a little clumsy, might be space for different APIs to access the data.

Started from `npm create vite@latest ref-browser -- --template react-ts`, which brought in a bunch of structure. Tried to minimize the dependencies and focus on functionality. There's plenty of room for improvement in the styling.

Had to make a few tweaks in `reference-lib`:

- ignore the `dist` folder in more tools
- rollup to an es module (the default) instead of common js. I was getting import errors with the `cjs` version, and our choice of `cjs` was copied from another example