ci: deploy to npm instead of github packages
Targets the public npm registry instead of the private github packages
to make this as easy to consume as possible.

Drop the dependency on a third party action for tagging, it's only a few lines to
add a tag, and the action was broken. Going with one less dependency to keep up with.