chore: update golang versions
Fix a broken build by upgrading to the latest golang version.

Also updated `gomarkdown` to resolve some low-priority security
notices and `golangci-lint` to keep up with the golang version.

The build broke because we're using "latest" versions of some dev
libraries. I considered pinning those instead of upgrading golang, but
"latest" fails infrequently, and we're already two minor versions
behind on golang so needed to update anyway.

If we have a similar failure will switch to a `tools.go` approach and
manage tool versions in go.mod with the rest of the dependencies.

fixes #307