update to go1.25, node24, golangci-lint v2, trixie
Considered pinning `gopls` again, but we want to be relatively up-to-date so figured I'd update things. I'm sure there's a better way to manage the node version, but this gets updated less than yearly so I'm not worried about it.

Made a few tweaks to satisfy golangci-lint but nothing significant.

fixes #482